+++ CHANGELOG.md
@@ -22,6 +22,8 @@ Security in case of vulnerabilities.
 
 ## Unreleased  
 ### Added  
+ - New route `/projects/team/unassigned`
+ - Your team projects "Unassigned" list (`/projects/team/unassigned`)
 
 ### Changed  
 - Sort all projects by region list alphabetically